requirements.txt
alembic~=1.12.0
asyncpg~=0.28.0
av~=11.0.0
boto3~=1.28.53
botocore~=1.31.57
certifi~=2023.7.22
coverage~=5.5
DateTime~=5.2
fastapi~=0.108.0
fastapi-users[sqlalchemy]
httpx~=0.25.0
Pillow~=10.0.1
psycopg2~=2.9.10
pydantic~=1.10.8
pytest-asyncio~=0.23.5
pytest~=8.0.0
python-dotenv~=1.0.0
python-magic-bin~=0.4.14
shotstack-sdk~=0.2.8
slowapi~=0.1.8
SQLAlchemy~=2.0.21
starlette~=0.32.0.post1
uuid~=1.30
uvicorn~=0.23.2